Create MinimumDrives.cs
Data consolidation program: Given a Integer() used, representing the amount of disk space used on each drive, and a corresponding Integer() total , representing the total capacity of each drive mentioned in used, we attempt to pack the data onto as few hard drives as possible. We may assume that the data consists of very small files, such that splitting it up and moving parts of it onto different hard drives never presents a problem. Returns the minimum number of hard drives that still contain data after the consolidation is complete.